band-archive/backend/requirements.txt

Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
python-dotenv==1.0.0
Werkzeug==3.0.1
pytest==7.4.3
